+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * fix(error): correctly parse in case of `RequestFailedError`
 * feat(run): add a `bash` alias for us out there often forgetting the `run` in front
 * feat(sshkeys): add support for ed25519 keys
+* feat(apps/create): detect Git main branch name
 
 ## 1.43.3